Update to Mattermost 5.28.1 (#94)
* Update to Mattermost 5.28.1

* Update release.mk patch for 5.28

* Repair to original patch text

* Run patch in batch mode

* Try to fix patch file

I'm having a hard time understanding why this is not working on CI.

* Working 5.28.1 release.mk patch

I was previously working off mattermost-server HEAD's copy of release.mk which has apparently changed since 5.28.1, so future PRs will need to make more changes to this patch. But for now, this should allow the patch to be applied.