Move config-reset to its own step
config-reset now (as of 5.36.0) is generated through a `go run` command.
If we pass the usual GOOS/GOARCH combo to this command, it will build
and then try to run an executable that's in a format different from the
host system, which can't succeed. This `go run`s config-reset as a
binary compatible with the host machine, then runs the full build with
the appropriate os/arch combo. All config-reset does is generate a new
config.json for the install anyway.

This make step was refactored in https://github.com/mattermost/mattermost-server/commit/3681cd3688408c96f1e26909a521bf72edceae57