Update release.mk patch
The Mattermost build script has changed to accomplish most of what we
were trying to patch it to support. The config generator is still a
point of contention, and I'm having trouble getting it to work even with
this patch, but wanted to toss it to the build machine to see if it
works there. We may need to dig into the config generator deeper if it
doesn't.

There are additional build script changes after 6.0 that have, for
example, linux-arm64 support already baked in, so we will eventually
need to make larger changes to which platforms we build at all.