Use node v15 by default
We are now getting this error when building:

    /build/.node-gyp/16.13.2/include/node/v8-internal.h:492:38: error: 'remove_cv_t' is not a member of 'std'

which the internet suggests is a problem with node 16 and node-sass. According to https://stackoverflow.com/questions/67241196/error-no-template-named-remove-cv-t-in-namespace-std-did-you-mean-remove , supposedly newer versions of node-sass support node 16, and the other solution is to pass -std=c++14 instead of c++0x (which is what is being done right now), but I can't say for sure whether this build script will work with CXXFLAGS or not. The other fix would seem to be to update node-sass to a version that supports node16, but that would presumably be more patching of files, this time in mattermost-webapp, so I'm trying the least-invasive option first.