Update how we retrieve the Mattermost version
With mmctl in the require list in the go.mod, mattermost-server no longer appears on the same line as the require statement, so splitting by spaces means we now need the second split entry instead of the third.

We could also fix this by removing mmctl from the go.mod, which I was going to do, but that means every Dependabot update would no longer work. So going forward we will need to ensure that the go.mod always has the mattermost-server require on its own line instead of inline with the require statement.