Use the mattermost-server version for mmctl
Since mmctl isn't versioned in a gomodules-friendly way, we either need to manually maintain the version in the release script or just assume/hope that they continue to create releases matching the mattermost-server version. So far that has always been the case, so we can ignore the go.mod version string.