Try Node 16 again
Last time we tried to use this, we ran into all sorts of errors due to node-sass, which is used by the mattermost frontend. Perhaps those issues have been resolved and we can move from 15 back to 16.

I'm trying this in the first place because our build logs show NPM throwing a fit about not supporting Node 15. And the last several PR builds have errored out in NPM in weird ways that the internet keeps suggesting "just update NPM" to fix, so here we are.

ref: commit 69dda8131efe815218fc8d26be3bd99ff2dd167f